// frontend/src/api/generated/model/error-code.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kafka Lab API
 * API for controlling and monitoring the Kafka lab demo.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * INVALID_MESSAGE_COUNT   – count doit être > 0 INVALID_PAYLOAD_SIZE    – payloadSizeBytes doit être > 0 PAYLOAD_TOO_LARGE       – payloadSizeBytes dépasse Integer.MAX_VALUE VALIDATION_ERROR        - fallback générique pour tout champ non mappé 
 */

export const ErrorCode = {
    InvalidMessageCount: 'INVALID_MESSAGE_COUNT',
    InvalidPayloadSize: 'INVALID_PAYLOAD_SIZE',
    PayloadTooLarge: 'PAYLOAD_TOO_LARGE',
    ValidationError: 'VALIDATION_ERROR',
} as const;

export type ErrorCode = typeof ErrorCode[keyof typeof ErrorCode];